requirements.txt
python-telegram-bot==21.10
reportlab==4.2.5
google-auth==2.38.0
google-auth-oauthlib==1.2.1
google-api-python-client==2.164.0
requests==2.32.3
